+++ dexs/fullsail-finance/index.ts
@@ -2,10 +2,11 @@ import fetchURL from "../../utils/fetchURL";
 import { FetchOptions, SimpleAdapter } from "../../adapters/types";
 import { CHAIN } from "../../helpers/chains";
 
-async function fetchSUI({ startTimestamp, endTimestamp }: FetchOptions) {
-    const dailyVolume = (await fetchURL(`https://app.fullsail.finance/api/defi_llama/volume?start_timestamp=${startTimestamp*1000}&end_timestamp=${endTimestamp*1000}`)).volume_usd;
+const fetch = async (options: FetchOptions) => {
+    const dailyVolume = (await fetchURL(`https://app.fullsail.finance/api/defi_llama/volume?start_timestamp=${options.startTimestamp * 1000}&end_timestamp=${options.endTimestamp * 1000}`)).volume_usd;
+
     return {
-        dailyVolume: dailyVolume,
+        dailyVolume,
     };
 }
 
@@ -13,7 +14,7 @@ const adapter: SimpleAdapter = {
     version: 2,
     adapter: {
         [CHAIN.SUI]: {
-            fetch: fetchSUI,
+            fetch,
             start: '2025-05-30',
         }
     }
