--- fees/FullSailFinance/index.ts
@@ -1,45 +1,0 @@
-import { SimpleAdapter, FetchResultV2, FetchV2 } from '../../adapters/types';
-import { CHAIN } from '../../helpers/chains';
-import fetchURL from '../../utils/fetchURL';
-
-const fullSailApiURL = 'https://app.fullsail.finance/api/defi_llama/fees';
-
-interface FullSailStats {
-    fees_usd: string;
-    protocol_revenue_usd: string;
-}
-
-const methodology = {
-    Fees: 'Swap fees generated by the swap transactions on FullSail finance protocol.',
-    ProtocolRevenue: 'Protocol fees charged from the swap fees.',
-};
-
-const fetchFullSailStats: FetchV2 = async ({
-    startTimestamp,
-    endTimestamp,
-}): Promise<FetchResultV2> => {
-    const url = `${fullSailApiURL}?start_timestamp=${startTimestamp*1000}&end_timestamp=${endTimestamp*1000}`;
-    const data: FullSailStats = await fetchURL(url);
-    const dailyFees = data.fees_usd;
-    const dailyRevenue = data.protocol_revenue_usd;
-    return {
-        dailyFees,
-        dailyRevenue,
-        dailyProtocolRevenue: dailyRevenue,
-    };
-};
-
-const adapter: SimpleAdapter = {
-    version: 2,
-    adapter: {
-        [CHAIN.SUI]: {
-            fetch: fetchFullSailStats,
-            start: '2025-05-30',
-            meta: {
-                methodology,
-            },
-        },
-    },
-};
-
-export default adapter;
